src/week_3/day_5_data_engineering/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
chromadb==0.5.0
click==8.1.7
debugpy==1.8.1
#ipython==8.24.0
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.46
langchain-text-splitters==0.0.1
langsmith==0.1.51
markdown-it-py==3.0.0
marshmallow==3.21.1
openai==1.23.6
pypdf==4.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
tokenizers==0.19.1
tomli==2.0.1
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.11.0
numpy==1.24.4
oauthlib==3.2.2
tiktoken